requirements.txt
https://github.com/New-dev0/Telethon/archive/Aurium.zip
https://github.com/kaif-00z/asyncpraw/archive/master.zip
python-dotenv
python-decouple
redis
aiohttp
